+++ src/app.ts
@@ -1,11 +1,13 @@
 import express from 'express';
 import dotenv from "dotenv";
+import authRouter from "./routes/auth.routes"
 
 dotenv.config();
 
 const app = express();
 //express json middleware
 app.use(express.json());
+app.use(authRouter)
 
 
 app.get('/health', (req, res)=>{