+++ src/config/config.ts
@@ -5,6 +5,18 @@ dotenv.config();
 const dburi = process.env.MONGO_URI;
 
 export async function connectDB() {
-	await mongoose.connect(process.env.MONGO_URI!);
-	console.log("connected to db");
+
+	if (!dburi) {
+        console.error("❌ Error: MONGO_URI is not defined in your environment variables!");
+        process.exit(1); 
+    }
+
+	try {
+        await mongoose.connect(dburi);
+        console.log("🍃 Connected to MongoDB successfully.");
+    } catch (error) {
+        console.error("❌ Database connection error:", error);
+        process.exit(1);
+    }
+
 }